saro-data-framework/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "saro-data-framework"
version = "1.1.0"
description = "Real-data testing framework for SARO — downloads public AI-risk datasets, converts them to SARO batch JSON, uploads, and validates all rules."
requires-python = ">=3.11"
license = { text = "MIT" }
dependencies = [
    "datasets>=3.2.0",
    "huggingface-hub>=0.27.1",
    "httpx>=0.28.1",
    "tenacity>=9.0.0",
    "tqdm>=4.67.1",
    "pydantic>=2.10.4",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "structlog>=24.4.0",
]

[project.scripts]
# saro_data package (new spec-compliant CLI)
saro-data = "saro_data.cli:cli"
# saro_data_framework package (legacy CLI kept for backward compat)
saro-data-legacy = "saro_data_framework.cli:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
# Both packages are installed from a single wheel
packages = [
    "src/saro_data",
    "src/saro_data_framework",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
